Bug fix: Due to a too small integer used in the function isotp_send_consecutive_frame(), frames larger than 255 bytes will not be sent correctly.
This bug was found by N. Bünder.

+++ isotp.c
@@ -106,7 +106,7 @@ static int isotp_send_first_frame(IsoTpLink* link, uint32_t id) {
 static int isotp_send_consecutive_frame(IsoTpLink* link) {
     
     IsoTpCanMessage message;
-    uint8_t data_length;
+    uint16_t data_length;
     int ret;
 
     /* multi frame message length must greater than 7  */
